Working version on Lightsail
This currently only runs on Lightsail by just using a single container
containing `mcm-app`, listening on port 5000 for HTTP.
I can't currently get it working with the nginx reverse proxy that works
locally. This is because `docker-compose` adds some DNS niceities so
that you can just redirect traffic to `app:5000`, where `app` is the
container name. However, this isn't setup on Lightsail so it fails.
I've tried:
  - localhost
  - 127.0.0.1
  - 0.0.0.0
  - <service>.service.local:<port> which should work according to the
    docs

Ideally I would have a solution that worked both locally and remotely,
so I can run `docker compose push` to push the images to ECR, then
restart the Lightsail service and the app has updated.