Can add species to mark list
Currently the only way to add a species to a mark list is via a button
on that species' own page. In future this will be expanded to allow for
the addition of species when browsing reactions, and also bulk selecting
of species groups as is currently done.

It might also be preferred to do this entirely in Javascript rather than
sending requests to the server constantly when Cookies can be extracted
in JS straight forwardly. The benefit of using Rack's Session variable
is that it allows us to port to other session managers, like DynamoDB