Updated bundler
There will still be 1 gem with a vulnerability failing audit: omniauth.
This is a dependency of uoy-faculty-sinatra-aws so I believe I need to
wait for this gem to update its Gemfile. The version of omniauth in the
base Sinatra template repo hasn't been updated to fix this vulnerability
as far as I can tell