Working tom select
The select box queries a new API after 300ms of no typing. It currently
finds any species or synonyms that match its criteria, and the results
are searched again and ranked by tom select.

This needs refining: the API currently just returns the Species name, so
Tom Select can't search for the synonym match that the API found. A
cleaner way of doing this is needed. Perhaps the API returns any Species
that has a match in either Formula or Synonym, along with a full list of
the synonyms for that compound. Tom Select then scores the search for
each hit according to whether the match was in the Formula or Synonym (I
_think_ searchField does this for us), and then again by how many
references the Synonym has.

However, this would involve searching twice. Perhaps it would be better
to do this entirely server side. I.e. find any Species that match by
formula or synonym then return for each Species: formula, synonyms (can
be truncated by TomSelect?), score (derived from whether match was in
formula or synonym)