Search working - results wrong order
TomSelect doesn't appear to be obeying the order in which the search
results are returned - having been ordered on the server end according
to specific criteria. For an example search for 'ethene' and observe
that C2H4 isn't the top result, despite being shown in the log to be the
highest scoring species.