Can add marklist from browse page
This is rather ugly as it stands, and adding values to the marklist
necessitates a page change as the request is POSTed and then redirected
back to the original page. Instead, it would be cleaner if this just
directly added to a cookie. This might mean using cookies directly in
Javascript rather than through Rack's session