Marklist has export button & is never auto-hidden
Hiding the marklist automatically, e.g. when the marklist was cleared,
was a little unfriendly to users both in terms of not necessarily being
the expected behaviour, but also the smooth transition animation wasn't
being applied. The marklist is still auto-shown when it is not visible
and a new species is added.