Don't display marklist on export page
This is an awkward fix, as the marklist is still being separately
populated in the HTML form, but it is now simply hidden from the user.
The reason why this has been done rather than simply remove the list
from this page is that the selected species are sent with the POST
request from this HTML form. If the form didn't exist, an XHR request
would need to be made instead, but these cannot automatically download
files by default, so it is cleaner from the code perspective to use the
HTML form, but simply not display it.