Mechanism select in navbar dropdown
While this functionality works, it could do with a touch-up for several
reasons. Firstly, there's an argument to be made that the MCM should be
first in the mechanism choices dropdown since it's the most important.
Secondly highlighting the current mechanism with the usual `.active`
class looks rather odd here, perhaps due to how the bootstrap class
'active' interacts with the custom header background. Furthermore, all
navbar items that use the `.nav-link` class are dulled.