Reactions ordered breadth-first in output
Also switched to Recursive CTE to traverse the mechanism rather than my
previous solution that involved heavy use of Ruby Sets. This is because
Sets are guaranteed to retain order, even though they were being used
for this reason. Given that the app may be updated to Ruby 3 soon, this
assumption could be broken.