Initial working export
The KPP export looks syntatically comparable to the old version,
but it needs testing in a model. One thing that looks different is the
old version appears to be combining 'identical' reactions (i.e. the same
reactants and products) by summing rates.

I can't see where in the codebase this is done and how valid it is, so I
need to investigate.